+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 This project uses the changelog in accordance with [keepchangelog](http://keepachangelog.com/). Please use this to write notable changes, which is not the same as git commit log...
 
 ## [unreleased][unreleased]
+- Changed `hf mf ecfill` to wait for execution and return status (@doegox)
 - Added option to wait for a card to `hf 14a reader` (@doegox)
 - Added support for quick dump via backdoor auth to `hf mf ecfill` (@doegox)
 - Fixed `hf mf restore` - really skip strict ACLs unless --force (@doegox)